requirements.txt
aiokafka==0.7.2
anyio==3.6.1
async-timeout==4.0.2
asyncpg==0.26.0
click==8.1.3
colorama==0.4.5
databases==0.6.1
Deprecated==1.2.13
fastapi==0.83.0
graphql-core==3.2.1
greenlet==1.1.3
h11==0.13.0
hiredis==2.0.0
httptools==0.5.0
idna==3.4
kafka-python==2.0.2
loguru==0.6.0
packaging==21.3
protobuf==4.21.6
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
redis==4.3.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
strawberry-graphql==0.130.3
typing_extensions==4.3.0
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
win32-setctime==1.1.0
wrapt==1.14.1
